@@ -6,7 +6,7 @@ package pkggraph
 import (
 	"bytes"
 	"fmt"
-	"io/ioutil"
+	"io"
 	"os"
 	"testing"
 
@@ -98,8 +98,7 @@ func TestMain(m *testing.M) {
 
 // buildRunNode creates a new 'Run' PkgNode based on a PackageVer struct
 func buildRunNodeHelper(pkg *pkgjson.PackageVer) (node *PkgNode) {
-	var pkgCopy pkgjson.PackageVer
-	pkgCopy = *pkg
+	pkgCopy := *pkg
 	node = &PkgNode{
 		VersionedPkg: &pkgCopy,
 		State:        StateMeta,
@@ -117,8 +116,7 @@ func buildRunNodeHelper(pkg *pkgjson.PackageVer) (node *PkgNode) {
 
 // buildBuildNode creates a new 'Build' PkgNode based on a PackageVer struct
 func buildBuildNodeHelper(pkg *pkgjson.PackageVer) (node *PkgNode) {
-	var pkgCopy pkgjson.PackageVer
-	pkgCopy = *pkg
+	pkgCopy := *pkg
 	node = &PkgNode{
 		VersionedPkg: &pkgCopy,
 		State:        StateBuild,
@@ -136,8 +134,7 @@ func buildBuildNodeHelper(pkg *pkgjson.PackageVer) (node *PkgNode) {
 
 // buildBuildNode creates a new 'Unresolved' PkgNode based on a PackageVer struct
 func buildUnresolvedNodeHelper(pkg *pkgjson.PackageVer) (node *PkgNode) {
-	var pkgCopy pkgjson.PackageVer
-	pkgCopy = *pkg
+	pkgCopy := *pkg
 	node = &PkgNode{
 		VersionedPkg: &pkgCopy,
 		State:        StateUnresolved,
@@ -166,6 +163,9 @@ func addNodeToGraphHelper(g *PkgGraph, node *PkgNode) (newNode *PkgNode, err err
 		node.Architecture,
 		node.SourceRepo,
 	)
+
+	// Updating node's ID for the sake of equality testing.
+	node.nodeID = newNode.nodeID
 	return
 }
 
@@ -225,7 +225,27 @@ func buildTestGraphHelper() (g *PkgGraph, err error) {
 	return
 }
 
+// Checks if two lists of nodes are equivalent (ignoring order and graph edges)
+func checkEqualComponents(t *testing.T, expected, actual []*PkgNode) {
+	t.Helper()
+	for _, mustHave := range expected {
+		foundPackage := false
+		for _, n := range actual {
+			foundPackage = foundPackage || mustHave.Equal(n)
+		}
+		assert.True(t, foundPackage, "expected to find %s in actual", mustHave.String())
+	}
+	for _, doHave := range actual {
+		foundPackage := false
+		for _, n := range expected {
+			foundPackage = foundPackage || doHave.Equal(n)
+		}
+		assert.True(t, foundPackage, "found %s in actual, but it was unexpected", doHave.String())
+	}
+}
+
 func checkTestGraph(t *testing.T, g *PkgGraph) {
+	t.Helper()
 	// Make sure we got the same graph back!
 	assert.Equal(t, len(allNodes), len(g.AllNodes()))
 	assert.Equal(t, len(runNodes)+len(unresolvedNodes), len(g.AllRunNodes()))
@@ -245,14 +265,7 @@ func checkTestGraph(t *testing.T, g *PkgGraph) {
 		pkgD2Unresolved,
 		pkgD3Unresolved,
 	}
-	for _, mustHave := range component1 {
-		found := false
-		for _, n := range g.AllNodesFrom(a.RunNode) {
-			found = found || mustHave.Equal(n)
-		}
-		assert.True(t, found)
-	}
-	assert.Equal(t, len(component1), len(g.AllNodesFrom(a.RunNode)))
+	checkEqualComponents(t, component1, g.AllNodesFrom(a.RunNode))
 
 	c2, err := g.FindBestPkgNode(&pkgjson.PackageVer{Name: "C"})
 	assert.NoError(t, err)
@@ -263,14 +276,7 @@ func checkTestGraph(t *testing.T, g *PkgGraph) {
 		pkgD5Unresolved,
 		pkgD6Unresolved,
 	}
-	for _, mustHave := range component2 {
-		found := false
-		for _, n := range g.AllNodesFrom(c2.RunNode) {
-			found = found || mustHave.Equal(n)
-		}
-		assert.True(t, found)
-	}
-	assert.Equal(t, len(component2), len(g.AllNodesFrom(c2.RunNode)))
+	checkEqualComponents(t, component2, g.AllNodesFrom(c2.RunNode))
 }
 
 // Validate the test graph is well formed
@@ -341,16 +347,23 @@ func TestDOTID(t *testing.T) {
 	for _, n := range allNodes {
 		assert.NotPanics(t, func() { n.DOTID() })
 	}
-	assert.Equal(t, "A-1-RUN<Meta> (ID=0,TYPE=Run,STATE=Meta)", pkgARun.DOTID())
-	assert.Equal(t, "D--REMOTE<Unresolved> (ID=0,TYPE=Remote,STATE=Unresolved)", pkgD1Unresolved.DOTID())
+
+	expectedARunDOTID := fmt.Sprintf("A-1-RUN<Meta> (ID=%d,TYPE=Run,STATE=Meta)", pkgARun.ID())
+	assert.Equal(t, expectedARunDOTID, pkgARun.DOTID())
+
+	expectedD1UnresolvedDOTID := fmt.Sprintf("D--REMOTE<Unresolved> (ID=%d,TYPE=Remote,STATE=Unresolved)", pkgD1Unresolved.ID())
+	assert.Equal(t, expectedD1UnresolvedDOTID, pkgD1Unresolved.DOTID())
 
 	g := NewPkgGraph()
 	goal, err := g.AddGoalNode("test", nil, nil, false)
 	assert.NoError(t, err)
-	assert.Equal(t, "test (ID=0,TYPE=Goal,STATE=Meta)", goal.DOTID())
+
+	expectedGoalDOTID := fmt.Sprintf("test (ID=%d,TYPE=Goal,STATE=Meta)", goal.ID())
+	assert.Equal(t, expectedGoalDOTID, goal.DOTID())
 
 	meta := g.AddMetaNode([]*PkgNode{}, []*PkgNode{})
-	assert.Equal(t, "Meta(1) (ID=1,TYPE=PureMeta,STATE=Meta)", meta.DOTID())
+	expectedMetaDOTID := fmt.Sprintf("Meta(1) (ID=%d,TYPE=PureMeta,STATE=Meta)", meta.ID())
+	assert.Equal(t, expectedMetaDOTID, meta.DOTID())
 
 	junk := PkgNode{State: -1, Type: -1}
 	assert.Panics(t, func() { junk.DOTID() })
@@ -358,10 +371,15 @@ func TestDOTID(t *testing.T) {
 
 // TestNodeString tests the built-in String() function for PkgNodes
 func TestNodeString(t *testing.T) {
-	assert.Equal(t, "A(1,):<ID:0 Type:Run State:Meta Rpm:A.rpm> from 'A.src.rpm' in 'test_repo'", pkgARun.String())
-	assert.Equal(t, "D(<1,):<ID:0 Type:Remote State:Unresolved Rpm:url://D.rpm> from 'url://D.src.rpm' in 'test_repo'", pkgD1Unresolved.String())
+	expectedARunString := fmt.Sprintf("A(1,):<ID:%d Type:Run State:Meta Rpm:A.rpm> from 'A.src.rpm' in 'test_repo'", pkgARun.ID())
+	assert.Equal(t, expectedARunString, pkgARun.String())
+
+	expectedD1UnresolvedString := fmt.Sprintf("D(<1,):<ID:%d Type:Remote State:Unresolved Rpm:url://D.rpm> from 'url://D.src.rpm' in 'test_repo'", pkgD1Unresolved.ID())
+	assert.Equal(t, expectedD1UnresolvedString, pkgD1Unresolved.String())
+
 	goalNode := PkgNode{GoalName: "goal", Type: TypeGoal, State: StateMeta}
 	assert.Equal(t, "goal():<ID:0 Type:Goal State:Meta Rpm:> from '' in ''", goalNode.String())
+
 	emptyNode := PkgNode{}
 	assert.Panics(t, func() { _ = emptyNode.String() })
 }
@@ -425,7 +443,6 @@ func TestAddMissingVersion(t *testing.T) {
 	n, err := addNodeToGraphHelper(g, pkgD2Unresolved)
 	assert.NoError(t, err)
 	assert.NotNil(t, n)
-
 }
 
 // Add a run node with an invalid version (for a run node)
@@ -723,8 +740,8 @@ func TestGoalWithPackages(t *testing.T) {
 	assert.Equal(t, len(runNodes)+len(unresolvedNodes), len(goalNodes))
 
 	goal, err = g.AddGoalNode("test2", []*pkgjson.PackageVer{
-		&pkgjson.PackageVer{Name: "A"},
-		&pkgjson.PackageVer{Name: "B"},
+		{Name: "A"},
+		{Name: "B"},
 	}, nil, false)
 	assert.NoError(t, err)
 	assert.NotNil(t, goal)
@@ -740,8 +757,95 @@ func TestStrictGoalNodes(t *testing.T) {
 	assert.NoError(t, err)
 	assert.NotNil(t, g)
 
-	_, err = g.AddGoalNode("test", []*pkgjson.PackageVer{&pkgjson.PackageVer{Name: "Not a package"}}, nil, true)
+	_, err = g.AddGoalNode("test", []*pkgjson.PackageVer{{Name: "Not a package"}}, nil, true)
 	assert.Error(t, err)
+}
+
+func TestGoalWithLevelZero(t *testing.T) {
+	g, err := buildTestGraphHelper()
+	assert.NoError(t, err)
+	assert.NotNil(t, g)
+
+	goal, err := g.AddGoalNode("test_0", []*pkgjson.PackageVer{&pkgC}, nil, false)
+	assert.NoError(t, err)
+	assert.NotNil(t, goal)
+	nodesInGoal := []*PkgNode{}
+	for _, n := range graph.NodesOf(g.From(goal.ID())) {
+		nodesInGoal = append(nodesInGoal, n.(*PkgNode))
+	}
+	expectedGoalNodes := []*PkgNode{
+		pkgCRun,
+	}
+	checkEqualComponents(t, expectedGoalNodes, nodesInGoal)
+	expectedGoalTree := []*PkgNode{
+		pkgCRun,
+		pkgCBuild,
+		pkgD3Unresolved,
+		goal,
+	}
+	checkEqualComponents(t, expectedGoalTree, g.AllNodesFrom(goal))
+}
+
+func TestGoalWithLevelOne(t *testing.T) {
+	g, err := buildTestGraphHelper()
+	assert.NoError(t, err)
+	assert.NotNil(t, g)
+
+	goal, err := g.AddGoalNodeWithExtraLayers("test_1", []*pkgjson.PackageVer{&pkgC}, nil, false, 1)
+	assert.NoError(t, err)
+	assert.NotNil(t, goal)
+	nodesInGoal := []*PkgNode{}
+	for _, n := range graph.NodesOf(g.From(goal.ID())) {
+		nodesInGoal = append(nodesInGoal, n.(*PkgNode))
+	}
+	expectedGoalNodes := []*PkgNode{
+		pkgCRun,
+		pkgBRun,
+	}
+	checkEqualComponents(t, expectedGoalNodes, nodesInGoal)
+	expectedGoalPackages := []*PkgNode{
+		pkgBRun,
+		pkgBBuild,
+		pkgCRun,
+		pkgCBuild,
+		pkgD2Unresolved,
+		pkgD3Unresolved,
+		goal,
+	}
+	checkEqualComponents(t, expectedGoalPackages, g.AllNodesFrom(goal))
+}
+
+func TestGoalWithLevelTwo(t *testing.T) {
+	g, err := buildTestGraphHelper()
+	assert.NoError(t, err)
+	assert.NotNil(t, g)
+
+	goal, err := g.AddGoalNodeWithExtraLayers("test_2", []*pkgjson.PackageVer{&pkgC}, nil, false, 2)
+	assert.NoError(t, err)
+	assert.NotNil(t, goal)
+	nodesInGoal := []*PkgNode{}
+	for _, n := range graph.NodesOf(g.From(goal.ID())) {
+		nodesInGoal = append(nodesInGoal, n.(*PkgNode))
+	}
+	expectedGoalNodes := []*PkgNode{
+		pkgARun,
+		pkgCRun,
+		pkgBRun,
+	}
+	checkEqualComponents(t, expectedGoalNodes, nodesInGoal)
+	expectedGoalPackages := []*PkgNode{
+		pkgARun,
+		pkgABuild,
+		pkgBRun,
+		pkgBBuild,
+		pkgCRun,
+		pkgCBuild,
+		pkgD1Unresolved,
+		pkgD2Unresolved,
+		pkgD3Unresolved,
+		goal,
+	}
+	checkEqualComponents(t, expectedGoalPackages, g.AllNodesFrom(goal))
 }
 
 // Add a meta node which should link the two disconnected graph components in the test graph
@@ -785,14 +889,7 @@ func TestMetaNode(t *testing.T) {
 		pkgD5Unresolved,
 		pkgD6Unresolved,
 	}
-	for _, mustHave := range component {
-		found := false
-		for _, n := range g.AllNodesFrom(a.RunNode) {
-			found = found || mustHave.Equal(n)
-		}
-		assert.True(t, found)
-	}
-	assert.Equal(t, len(component), len(g.AllNodesFrom(a.RunNode)))
+	checkEqualComponents(t, component, g.AllNodesFrom(a.RunNode))
 }
 
 // Make sure the graph updates after adding meta nodes
@@ -822,14 +919,7 @@ func TestMetaNodeAddPkg(t *testing.T) {
 		pkgD5Unresolved,
 		pkgD6Unresolved,
 	}
-	for _, mustHave := range component {
-		found := false
-		for _, n := range g.AllNodesFrom(a.RunNode) {
-			found = found || mustHave.Equal(n)
-		}
-		assert.True(t, found)
-	}
-	assert.Equal(t, len(component), len(g.AllNodesFrom(a.RunNode)))
+	checkEqualComponents(t, component, g.AllNodesFrom(a.RunNode))
 
 	n, err := addNodeToGraphHelper(g, buildUnresolvedNodeHelper(&pkgjson.PackageVer{Name: "test", Version: "99"}))
 	assert.NoError(t, err)
@@ -839,6 +929,86 @@ func TestMetaNodeAddPkg(t *testing.T) {
 	assert.NoError(t, err)
 	assert.Equal(t, 9+5+1+1, len(g.AllNodesFrom(a.RunNode)))
 	assert.Equal(t, 5, len(g.AllNodesFrom(c.RunNode)))
+}
+
+func TestGoalWithLevelOneAndMeta(t *testing.T) {
+	g, err := buildTestGraphHelper()
+	assert.NoError(t, err)
+	assert.NotNil(t, g)
+
+	c1, err := g.FindBestPkgNode(&pkgC)
+	assert.NoError(t, err)
+	c2, err := g.FindBestPkgNode(&pkgC2)
+	assert.NoError(t, err)
+	meta := g.AddMetaNode([]*PkgNode{c2.RunNode}, []*PkgNode{c1.RunNode})
+
+	goal, err := g.AddGoalNodeWithExtraLayers("test_1meta", []*pkgjson.PackageVer{&pkgC}, nil, false, 1)
+	assert.NoError(t, err)
+	assert.NotNil(t, goal)
+	nodesInGoal := []*PkgNode{}
+	for _, n := range graph.NodesOf(g.From(goal.ID())) {
+		nodesInGoal = append(nodesInGoal, n.(*PkgNode))
+	}
+	expectedGoalNodes := []*PkgNode{
+		pkgCRun,
+		pkgC2Run,
+		pkgBRun,
+	}
+	checkEqualComponents(t, expectedGoalNodes, nodesInGoal)
+	// But we now pull in the entire graph when looking at the tree
+	expectedGoalPackagesMeta := []*PkgNode{
+		pkgBRun,
+		pkgBBuild,
+		pkgCRun,
+		pkgCBuild,
+		pkgD2Unresolved,
+		pkgD3Unresolved,
+		pkgC2Run,
+		pkgC2Build,
+		pkgD4Unresolved,
+		pkgD5Unresolved,
+		pkgD6Unresolved,
+		meta,
+		goal,
+	}
+	checkEqualComponents(t, expectedGoalPackagesMeta, g.AllNodesFrom(goal))
+}
+
+func TestGoalWithMultipleGoalsAndOneExtraLayer(t *testing.T) {
+	g, err := buildTestGraphHelper()
+	assert.NoError(t, err)
+	assert.NotNil(t, g)
+
+	goal, err := g.AddGoalNodeWithExtraLayers("test_1multi", []*pkgjson.PackageVer{&pkgC, &pkgD4}, nil, false, 1)
+	assert.NoError(t, err)
+	assert.NotNil(t, goal)
+	nodesInGoal := []*PkgNode{}
+	for _, n := range graph.NodesOf(g.From(goal.ID())) {
+		nodesInGoal = append(nodesInGoal, n.(*PkgNode))
+	}
+	expectedGoalNodes := []*PkgNode{
+		pkgCRun,
+		pkgC2Run,
+		pkgD4Unresolved,
+		pkgBRun,
+	}
+	checkEqualComponents(t, expectedGoalNodes, nodesInGoal)
+	// But we now pull in the entire graph when looking at the tree
+	expectedGoalPackagesMeta := []*PkgNode{
+		pkgBRun,
+		pkgBBuild,
+		pkgCRun,
+		pkgCBuild,
+		pkgD2Unresolved,
+		pkgD3Unresolved,
+		pkgC2Run,
+		pkgC2Build,
+		pkgD4Unresolved,
+		pkgD5Unresolved,
+		pkgD6Unresolved,
+		goal,
+	}
+	checkEqualComponents(t, expectedGoalPackagesMeta, g.AllNodesFrom(goal))
 }
 
 // Test encoding and decoding a DOT formatted graph
@@ -891,6 +1061,7 @@ func TestEncodeDecodeMultiDOT(t *testing.T) {
 
 	gFinal := NewPkgGraph()
 	err = ReadDOTGraph(gFinal, &buf2)
+	assert.NoError(t, err)
 
 	checkTestGraph(t, gFinal)
 }
@@ -918,6 +1089,13 @@ func TestReadWriteGraph(t *testing.T) {
 
 // Validate the reference graph is valid, and that it matches the output of the test graph.
 func TestReferenceDOTFile(t *testing.T) {
+	if testing.Short() {
+		// Encoding is unreliable on some systems. The final output is still a valid
+		// graph but the base64 encoding of the nodes is different. The final graph once
+		// decoded is the same however (probably gob encoding is different since its stateful?)
+		t.Skip("Short mode enabled")
+	}
+
 	gIn, err := ReadDOTGraphFile("test_graph_reference.dot")
 	assert.NoError(t, err)
 
@@ -932,13 +1110,15 @@ func TestReferenceDOTFile(t *testing.T) {
 	assert.NoError(t, err)
 
 	f, err := os.Open("test_graph_reference.dot")
-	defer f.Close()
+	if err == nil {
+		defer f.Close()
+	}
 	assert.NoError(t, err)
 
 	// Compare the bytes from the reference file against a fresh encoding
-	bytesFromCode, err := ioutil.ReadAll(&buf)
+	bytesFromCode, err := io.ReadAll(&buf)
 	assert.NoError(t, err)
-	bytesFromFile, err := ioutil.ReadAll(f)
+	bytesFromFile, err := io.ReadAll(f)
 	assert.NoError(t, err)
 	assert.True(t, len(bytesFromCode) > 0)
 	assert.True(t, len(bytesFromFile) > 0)
@@ -990,6 +1170,7 @@ func TestEncodingSubGraph(t *testing.T) {
 
 	// Copy uses the encode/decode flow
 	gCopy, err := subGraph.DeepCopy()
+	assert.NoError(t, err)
 
 	component := []*PkgNode{
 		pkgCRun,
@@ -1068,4 +1249,48 @@ func TestShouldGetSRPMNameFromEmptySRPMPath(t *testing.T) {
 	}
 
 	assert.Equal(t, ".", node.SRPMFileName())
+}
+
+func TestShouldGetAllBuildNodesWithFilter(t *testing.T) {
+	gOut, err := buildTestGraphHelper()
+	assert.NoError(t, err)
+	assert.NotNil(t, gOut)
+
+	foundNodes := gOut.NodesMatchingFilter(func(node *PkgNode) bool {
+		return node.Type == TypeLocalBuild
+	})
+	checkEqualComponents(t, buildNodes, foundNodes)
+}
+
+func TestShouldGetAllNodesWithFilter(t *testing.T) {
+	gOut, err := buildTestGraphHelper()
+	assert.NoError(t, err)
+	assert.NotNil(t, gOut)
+
+	foundNodes := gOut.NodesMatchingFilter(func(node *PkgNode) bool {
+		return true
+	})
+	checkEqualComponents(t, allNodes, foundNodes)
+}
+
+func TestShouldGetAllRunNodesWithFilter(t *testing.T) {
+	gOut, err := buildTestGraphHelper()
+	assert.NoError(t, err)
+	assert.NotNil(t, gOut)
+
+	foundNodes := gOut.NodesMatchingFilter(func(node *PkgNode) bool {
+		return node.Type == TypeLocalRun
+	})
+	checkEqualComponents(t, runNodes, foundNodes)
+}
+
+func TestShouldGetAllUnresolvedNodesWithFilter(t *testing.T) {
+	gOut, err := buildTestGraphHelper()
+	assert.NoError(t, err)
+	assert.NotNil(t, gOut)
+
+	foundNodes := gOut.NodesMatchingFilter(func(node *PkgNode) bool {
+		return node.State == StateUnresolved
+	})
+	checkEqualComponents(t, unresolvedNodes, foundNodes)
 }
