@@ -13,6 +13,7 @@ import (
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/file"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/logger"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/pkggraph"
+	"github.com/microsoft/CBL-Mariner/toolkit/tools/internal/sliceutils"
 	"github.com/microsoft/CBL-Mariner/toolkit/tools/scheduler/schedulerutils"
 
 	"gonum.org/v1/gonum/graph"
@@ -89,7 +90,7 @@ func main() {
 	nodeListGoal := searchForGoal(graph, goalSearchList)
 
 	nodeLists := append(nodeListPkg, append(nodeListSpec, nodeListGoal...)...)
-	nodeSet := removeDuplicates(nodeLists)
+	nodeSet := sliceutils.RemoveDuplicatesFromSlice(nodeLists)
 
 	if len(nodeSet) == 0 {
 		logger.Log.Panicf("Could not find any nodes matching pkgs:[%s] or specs:[%s] or goals[%s]", *pkgsToSearch, *specsToSearch, *goalsToSearch)
@@ -172,18 +173,6 @@ func searchForSpec(graph *pkggraph.PkgGraph, specs []string) (list []*pkggraph.P
 	return
 }
 
-func removeDuplicates(nodeList []*pkggraph.PkgNode) (uniqueNodeList []*pkggraph.PkgNode) {
-	nodeMap := make(map[*pkggraph.PkgNode]bool)
-	for _, n := range nodeList {
-		nodeMap[n] = true
-	}
-	uniqueNodeList = make([]*pkggraph.PkgNode, 0, len(nodeMap))
-	for key, _ := range nodeMap {
-		uniqueNodeList = append(uniqueNodeList, key)
-	}
-	return
-}
-
 func buildRequiresGraph(graphIn *pkggraph.PkgGraph, nodeList []*pkggraph.PkgNode) (graphOut *pkggraph.PkgGraph, root *pkggraph.PkgNode, err error) {
 	// Make a copy of the graph
 	newGraph, err := graphIn.DeepCopy()
@@ -242,7 +231,7 @@ func formatNode(n *pkggraph.PkgNode, verbosity int) string {
 	case 3:
 		return fmt.Sprintf("'%s' from node '%s'", filepath.Base(n.RpmPath), n.FriendlyName())
 	case 4:
-		return fmt.Sprintf("'%s'", n)
+		return fmt.Sprintf("(%v)'%#v'", n.VersionedPkg, *n)
 	default:
 		logger.Log.Fatalf("Invalid verbosity level %v", verbosity)
 	}
@@ -256,10 +245,7 @@ func isFilteredFile(path, filterFile string) bool {
 			if err != nil {
 				logger.Log.Fatalf("Failed to load filter file '%s': %s", filterFile, err)
 			}
-			reservedFiles = make(map[string]bool)
-			for _, f := range reservedFileList {
-				reservedFiles[f] = true
-			}
+			reservedFiles = sliceutils.SliceToSet[string](reservedFileList)
 		}
 		base := filepath.Base(path)
 
@@ -449,10 +435,7 @@ func printSpecs(graph *pkggraph.PkgGraph, tree, filter bool, filterFile string, 
 		}
 
 		// Contert to list and sort
-		printLines := []string{}
-		for s := range results {
-			printLines = append(printLines, s)
-		}
+		printLines := sliceutils.SetToSlice[string](results)
 		sort.Strings(printLines)
 		for _, l := range printLines {
 			fmt.Println(l)
