@@ -156,7 +156,12 @@ func cloneSystemConfigs(cloner repocloner.RepoCloner, configFile, baseDirPath st
 
 	logger.Log.Infof("Cloning: %v", packageVersionsInConfig)
 	// The image tools don't care if a package was created locally or not, just that it exists. Disregard if it is prebuilt or not.
-	_, err = cloner.Clone(cloneDeps, packageVersionsInConfig...)
+	_, err = cloner.CloneByPackageVerSingleTransaction(cloneDeps, packageVersionsInConfig...)
+	if err != nil {
+		// Fallback to legacy flow with multiple transactions in case we get a OOM error from a large transaction.
+		logger.Log.Warnf("Failed to clone packages in a single transaction, will retry with individual transactions... (%s)", err)
+		_, err = cloner.CloneByPackageVer(cloneDeps, packageVersionsInConfig...)
+	}
 	return
 }
 
