@@ -3,7 +3,10 @@
 
 package buildagents
 
-import "fmt"
+import (
+	"fmt"
+	"time"
+)
 
 // BuildAgentConfig represents configuration options a BuildAgent would need to successfully build a given package.
 type BuildAgentConfig struct {
@@ -17,6 +20,7 @@ type BuildAgentConfig struct {
 	SrpmDir      string
 	CacheDir     string
 	CCacheDir    string
+	CCacheConfig string
 
 	DistTag              string
 	DistroReleaseVersion string
@@ -26,6 +30,7 @@ type BuildAgentConfig struct {
 	NoCleanup bool
 	UseCcache bool
 	MaxCpu    string
+	Timeout   time.Duration
 
 	LogDir   string
 	LogLevel string
@@ -37,12 +42,13 @@ type BuildAgent interface {
 	Initialize(config *BuildAgentConfig) error
 
 	// BuildPackage builds a given file and returns the output files or error.
+	// - basePackageName is the base package name derived from the spec file (i.e. 'kernel').
 	// - inputFile is the SRPM to build.
 	// - logName is the file name to save the package build log to.
 	// - outArch is the target architecture to build for.
 	// - runCheck is true if the package should run the "%check" section during the build
 	// - dependencies is a list of dependencies that need to be installed before building.
-	BuildPackage(inputFile, logName, outArch string, runCheck bool, dependencies []string) ([]string, string, error)
+	BuildPackage(basePackageName, inputFile, logName, outArch string, runCheck bool, dependencies []string) ([]string, string, error)
 
 	// Config returns a copy of the agent's configuration.
 	Config() BuildAgentConfig
